# src/config/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.1.0",
    "title": "API Documentation",
    "description": "API Documentation"
  },
  "host": "localhost:3002",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/admin/mon/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "IDMon": {
                  "example": "any"
                },
                "Image": {
                  "example": "any"
                },
                "DonGiaBanSi": {
                  "example": "any"
                },
                "DonGiaBanLe": {
                  "example": "any"
                },
                "DonGiaVon": {
                  "example": "any"
                },
                "SoLuongTonKho": {
                  "example": "any"
                },
                "ThoiGianBH": {
                  "example": "any"
                },
                "TenKhongDau": {
                  "example": "any"
                },
                "TenMon": {
                  "example": "any"
                },
                "IDLoaiMon": {
                  "example": "any"
                },
                "createDate": {
                  "example": "any"
                },
                "createBy": {
                  "example": "any"
                },
                "modifyDate": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/mon/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "Image": {
                  "example": "any"
                },
                "TenKhongDau": {
                  "example": "any"
                },
                "TenMon": {
                  "example": "any"
                },
                "modifyDate": {
                  "example": "any"
                },
                "modifyBy": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}